new file:   Check_Installed_Software.ps1
	new file:   Delete_Shortcut_w-URL.ps1
	modified:   Grab_Logs.ps1
	new file:   Remove_ManageEnigine.ps1
	modified:   Replace_Old_Shortcut.ps1

+++ Grab_Logs.ps1
@@ -1,3 +1,20 @@
+# Function to check if we are running as administrator
+function Test-IsAdmin {
+    $identity = [Security.Principal.WindowsIdentity]::GetCurrent()
+    $principal = New-Object Security.Principal.WindowsPrincipal $identity
+    $principal.IsInRole([Security.Principal.WindowsBuiltInRole]::Administrator)
+}
+
+# Check if we are not running as an administrator
+if (-not (Test-IsAdmin)) {
+    # Relaunch the script with elevated rights
+    Start-Process powershell -ArgumentList "-NoProfile -ExecutionPolicy Bypass -File $($MyInvocation.MyCommand.Definition)" -Verb RunAs
+    return
+}
+
+# Load Windows Forms assembly
+Add-Type -AssemblyName System.Windows.Forms
+
 # Get all the log names
 $logNames = Get-WinEvent -ListLog * | Select-Object -ExpandProperty LogName
 
@@ -20,11 +37,23 @@ foreach ($logName in $logNames) {
     }
 }
 
-# Compress the logs into a ZIP file
-$zipPath = "$($newFolder.FullName).zip"
-Compress-Archive -Path "$newFolder\*" -DestinationPath $zipPath -CompressionLevel Optimal
+# Ask user for save location using Save File Dialog
+$saveFileDialog = New-Object System.Windows.Forms.SaveFileDialog
+$saveFileDialog.Filter = "ZIP files (*.zip)|*.zip"
+$saveFileDialog.Title = "Save the logs ZIP file"
+$saveFileDialog.ShowDialog() | Out-Null
+$zipPath = $saveFileDialog.FileName
 
-# Delete the uncompressed logs
-Remove-Item -Recurse -Force $newFolder
+# Ensure that a path was provided
+if (-not [string]::IsNullOrWhiteSpace($zipPath)) {
+    # Compress the logs into a ZIP file
+    Compress-Archive -Path "$newFolder\*" -DestinationPath $zipPath -Force -CompressionLevel Optimal
 
-Write-Host "Finished exporting and compressing logs to $zipPath."
+    # Delete the uncompressed logs
+    Remove-Item -Recurse -Force $newFolder
+
+    Write-Host "Finished exporting and compressing logs to $zipPath."
+}
+else {
+    Write-Warning "No save location selected. Aborting compression."
+}
